refactor: move Collection and Collector class to Tracking domain
